refactor: Race 객체에서 사용자 입력은 Controller로 분리
