+++ src/main/java/racingcar/Application.java
@@ -1,8 +1,10 @@
 package racingcar;
 
+import racingcar.controller.RaceController;
+
 public class Application {
     public static void main(String[] args) {
-        Race race = new Race();
-        race.open();
+        RaceController raceController = new RaceController();
+        raceController.raceStart();
     }
 }
